requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.1
idna==3.11
numpy==2.4.1
pandas==2.3.3
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
